Replacing `createFile` with `Data.write`
This, coupled with the removal of the `fileExists` check, prevents a race condition witnessed on macOS with optimized code.

We can safely remove the `fileExists` check because we are now using the `.withoutOverwriting` option which will cause the `write` to fail if the file exists. This approach is safer since we run the potential for a race condition between the time of the `fileExists` check and the creation of the file, in the original implementation.