Add product with dynamic link type
Hi, I would like to add the possibility to import the library with dynamic link type. I have a XCFramework which has the ZIPFoundation as its dependency, however I am not able to link the ZIPFoundation because SPM imports it as static library so at runtime my app crashes because  dyld Symbol for ZIPFoundation was not found. By using the product ZIPFoundationDynamic with dynamic link I was able to fix the issue.